Moderation backend ready (Not tested)

+++ sts/src/main/java/util/IWFileManager.java
@@ -6,18 +6,22 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Paths;
 
+import javax.persistence.EntityManager;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.multipart.MultipartFile;
 
 import es.fdi.iw.ContextInitializer;
+import es.fdi.iw.model.User;
 
 public class IWFileManager 
 {
 	private File baseDirectory;
+	private EntityManager manager;
 	static final Logger log = LoggerFactory.getLogger(IWFileManager.class);
 	
-	public IWFileManager(String basePath)
+	public IWFileManager(String basePath, EntityManager manager)
 	{
 		baseDirectory = new File(basePath);
 		log.info("base directory is {}", baseDirectory.getAbsolutePath());
@@ -29,6 +33,8 @@ public class IWFileManager
     			log.error("{} could not be created -- check permissions", baseDirectory);        			
     	else
     		log.info("using already-existing base folder :-)");
+		
+		this.manager = manager;
 	}
 	
 	public String getFilePath(es.fdi.iw.model.File file)
@@ -40,14 +46,22 @@ public class IWFileManager
 	{
 		File f = new File(getFilePath(file));
 		
-		f.createNewFile(); //See the docs, just returns false if the file already exists
+		f.createNewFile(); //See the docs, just returns false if the file already exists	
 		
 		return f;
 	}
 	
-	public void deleteFile(es.fdi.iw.model.File file) throws IOException
+	public void deleteFile(es.fdi.iw.model.File file) 
 	{
-		getFile(file).delete();
+		//An user has no direct rights to delete a file, just queue this event
+		IWModerationManager.get(manager).moderateDeleteFile(file);
+	}
+	
+	public void trueDeleteFile(es.fdi.iw.model.File file, User moderator) throws IOException {
+		if(moderator.getRole() == "moderator") {
+			getFile(file).delete();
+			manager.remove(file);
+		}
 	}
 	
 	public es.fdi.iw.model.File uploadFile(MultipartFile load, String tags)
@@ -62,6 +76,10 @@ public class IWFileManager
                         		new FileOutputStream(getFile(file)));
                 stream.write(bytes);
                 stream.close();
+                
+                manager.persist(file);
+                
+                IWModerationManager.get(manager).moderateNewFile(file);
                 
                 return file;
             } catch (Exception e) {
